+++ src/env/node.ts
@@ -1,16 +1,14 @@
 /* eslint-disable no-restricted-globals */
 import type { Canvas as NodeCanvas } from 'canvas';
 import { JSDOM } from 'jsdom';
-// @ts-ignore
+// @ts-expect-error internal import
 import utils from 'jsdom/lib/jsdom/living/generated/utils.js';
 import { config } from '../config';
 import { NodeGLProbe } from '../filters/GLProbes/NodeGLProbe';
 import { setEnv } from './index';
-import type { TCopyPasteData, TFabricEnv } from './types';
+import type { TFabricEnv } from './types';
 
 const { implForWrapper: jsdomImplForWrapper } = utils;
-
-const copyPasteData: TCopyPasteData = {};
 
 const { window: JSDOMWindow } = new JSDOM(
   decodeURIComponent(
@@ -51,7 +49,6 @@ export const getEnv = (): TFabricEnv => {
     isTouchSupported: false,
     WebGLProbe: new NodeGLProbe(),
     dispose,
-    copyPasteData,
   };
 };