+++ test/visual/text.js
@@ -584,6 +584,9 @@
         },
         setDragImage(imageSource, x, y) {
 
+        },
+        clearData() {
+          
         }
       }
     });